beamserch implementation in refactored optim (3)

+++ plugins/optim/src/strategy/ibex_CellBeamSearch.cpp
@@ -14,7 +14,7 @@ using namespace std;
 
 namespace ibex {
 
-  const double CellBeamSearch::default_beamsize= 1;
+  const unsigned int  CellBeamSearch::default_beamsize= 1;
 
 
   CellBeamSearch::CellBeamSearch (CellHeap& currentbuffer, CellHeap& futurebuffer, const ExtendedSystem & sys, unsigned int beamsize) : CellHeap (sys) , currentbuffer(currentbuffer), futurebuffer (futurebuffer), beamsize(beamsize) {}
